search algos

+++ 0x1E-search_algorithms/0-linear.c
@@ -6,11 +6,12 @@
  * @array: array to be searched
  * @size: size of array
  * @value: value being searched for
+ *
  * Return: Always EXIT_SUCCESS
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL)
 		return (-1);
